+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/block/custom/BigSom.java
@@ -14,20 +14,27 @@
  *
  */
 
-package org.scilab.modules.xcos.block.positionning;
+package org.scilab.modules.xcos.block.custom;
 
 import com.mxgraph.model.mxGeometry;
 import org.scilab.modules.xcos.JavaController;
 import org.scilab.modules.xcos.Kind;
 import org.scilab.modules.xcos.block.BasicBlock;
+import org.scilab.modules.xcos.block.custom.SumPortLabeler;
 
 /**
- * Implement the PRODUCT_f block.
+ * Implement the BIGSOM_f function.
  */
-@SuppressWarnings(value = { "serial" })
-public class Product extends BasicBlock {
-
-    public Product(JavaController controller, long uid, Kind kind, Object value, mxGeometry geometry, String style, String id) {
+public class BigSom extends BasicBlock {
+    /**
+     * Default constructor
+     */
+    public BigSom(JavaController controller, long uid, Kind kind, Object value, mxGeometry geometry, String style, String id) {
         super(controller, uid, kind, value, geometry, style, id);
+    }
+
+    @Override
+    public void updateBlockView() {
+        SumPortLabeler.getInstance().updateLabel(this);
     }
 }