+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/block/custom/SumPortLabeler.java
@@ -1,7 +1,7 @@
 /*
  * Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
  * Copyright (C) 2010 - DIGITEO - Clement DAVID
- * Copyright (C) 2011-2015 - Scilab Enterprises - Clement DAVID
+ * Copyright (C) 2011-2016 - Scilab Enterprises - Clement DAVID
  *
  * Copyright (C) 2012 - 2016 - Scilab Enterprises
  *
@@ -14,44 +14,42 @@
  *
  */
 
-package org.scilab.modules.xcos.block.listener;
+package org.scilab.modules.xcos.block.custom;
 
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.scilab.modules.types.ScilabDouble;
 import org.scilab.modules.xcos.block.BasicBlock;
 import org.scilab.modules.xcos.port.input.InputPort;
 
 import com.mxgraph.model.mxICell;
+import org.scilab.modules.xcos.JavaController;
+import org.scilab.modules.xcos.Kind;
+import org.scilab.modules.xcos.ObjectProperties;
+import org.scilab.modules.xcos.VectorOfDouble;
 
 /**
  * Change the port label on rpar change.
  *
  * This listener must be installed on "integerParameters" property.
  */
-@SuppressWarnings(value = { "serial" })
-public class SumPortLabelingListener implements PropertyChangeListener,
-    Serializable {
+public class SumPortLabeler {
 
     private static final String NOT_PRINTED_LABEL = "+";
-    private static SumPortLabelingListener instance;
+    private static SumPortLabeler instance;
 
     /**
      * Default constructor
      */
-    public SumPortLabelingListener() {
+    public SumPortLabeler() {
     }
 
     /**
      * @return the shared instance
      */
-    public static SumPortLabelingListener getInstance() {
+    public static SumPortLabeler getInstance() {
         if (instance == null) {
-            instance = new SumPortLabelingListener();
+            instance = new SumPortLabeler();
         }
         return instance;
     }
@@ -60,15 +58,9 @@ public class SumPortLabelingListener implements PropertyChangeListener,
      * Change the label of the port according to the integer parameters
      * property.
      *
-     * @param evt
-     *            the event
-     * @see java.beans.PropertyChangeListener#propertyChange(java.beans.PropertyChangeEvent)
+     * @param source the block to update
      */
-    @Override
-    public void propertyChange(PropertyChangeEvent evt) {
-        final BasicBlock source = (BasicBlock) evt.getSource();
-        final ScilabDouble data = (ScilabDouble) evt.getNewValue();
-
+    public void updateLabel(BasicBlock source) {
         /**
          * Get the input port children
          */
@@ -84,19 +76,21 @@ public class SumPortLabelingListener implements PropertyChangeListener,
         /**
          * Set the ports labels
          */
-        // FIXME
-        //        for (InputPort port : ports) {
-        //            final double gain;
-        //
-        //            if (data.isEmpty()
-        //                    || data.getRealPart().length < port.getOrdering()) {
-        //                gain = 1;
-        //            } else {
-        //                gain = data.getRealPart()[port.getOrdering() - 1][0];
-        //            }
-        //
-        //            port.setValue(getLabel(gain));
-        //        }
+        JavaController controller = new JavaController();
+        VectorOfDouble rpar = new VectorOfDouble(ports.size());
+        controller.getObjectProperty(source.getUID(), Kind.BLOCK, ObjectProperties.RPAR, rpar);
+
+        for (int i = 0; i < ports.size(); i++) {
+            final double gain;
+
+            if (rpar.size() < i) {
+                gain = 1;
+            } else {
+                gain = rpar.get(i);
+            }
+
+            ports.get(i).setValue(getLabel(gain));
+        }
 
         /**
          * Check if all the values are equal to the default one.